+++ lib/utils/constants.ts
@@ -1,5 +1,0 @@
-import type { ClearOptions } from '@/utils/types'
-
-export const DEFAULT_CLEAR_OPTIONS: ClearOptions = {
-  clearAuto: true
-}